+++ src/names.c
@@ -10,6 +10,42 @@ static void describe_repair(SEXP old_names, SEXP new_names);
 // 3 leading '.' + 1 trailing '\0' + 24 characters
 #define MAX_IOTA_SIZE 28
 
+// Initialised at load time
+SEXP syms_as_universal_names = NULL;
+SEXP syms_validate_unique_names = NULL;
+SEXP fns_as_universal_names = NULL;
+SEXP fns_validate_unique_names = NULL;
+
+// Defined below
+SEXP vctrs_as_minimal_names(SEXP names);
+SEXP vec_as_universal_names(SEXP names, bool quiet);
+SEXP vec_validate_unique_names(SEXP names);
+
+
+// [[ include("vctrs.h") ]]
+SEXP vec_as_names(SEXP names, enum name_repair_arg type, bool quiet) {
+  switch (type) {
+  case name_repair_none: return names;
+  case name_repair_minimal: return vctrs_as_minimal_names(names);
+  case name_repair_unique: return vec_as_unique_names(names, quiet);
+  case name_repair_universal: return vec_as_universal_names(names, quiet);
+  case name_repair_check_unique: return vec_validate_unique_names(names);
+  }
+}
+
+SEXP vec_as_universal_names(SEXP names, bool quiet) {
+  SEXP quiet_obj = PROTECT(r_lgl(quiet));
+  SEXP out = vctrs_dispatch2(syms_as_universal_names, fns_as_universal_names,
+                             syms_names, names,
+                             syms_quiet, quiet_obj);
+  UNPROTECT(1);
+  return out;
+}
+SEXP vec_validate_unique_names(SEXP names) {
+  return vctrs_dispatch1(syms_validate_unique_names, fns_validate_unique_names,
+                         syms_names, names);
+}
+
 
 // [[ register(); include("vctrs.h") ]]
 SEXP vec_names(SEXP x) {
@@ -36,6 +72,7 @@ SEXP vec_names(SEXP x) {
   return out;
 }
 
+// [[ register() ]]
 SEXP vctrs_as_minimal_names(SEXP names) {
   if (TYPEOF(names) != STRSXP) {
     Rf_errorcall(R_NilValue, "`names` must be a character vector");
@@ -68,6 +105,7 @@ SEXP vctrs_as_minimal_names(SEXP names) {
   return names;
 }
 
+// [[ register() ]]
 SEXP vctrs_minimal_names(SEXP x) {
   SEXP names = PROTECT(vec_names(x));
 
@@ -93,7 +131,7 @@ static ptrdiff_t suffix_pos(const char* name);
 static bool needs_suffix(SEXP str);
 
 // [[ include("vctrs.h") ]]
-SEXP as_unique_names(SEXP names, bool quiet) {
+SEXP vec_as_unique_names(SEXP names, bool quiet) {
   if (is_unique_names(names) && !any_has_suffix(names)) {
     return names;
   } else {
@@ -208,7 +246,7 @@ SEXP as_unique_names_impl(SEXP names, bool quiet) {
 }
 
 SEXP vctrs_as_unique_names(SEXP names, SEXP quiet) {
-  SEXP out = PROTECT(as_unique_names(names, LOGICAL(quiet)[0]));
+  SEXP out = PROTECT(vec_as_unique_names(names, LOGICAL(quiet)[0]));
   UNPROTECT(1);
   return out;
 }
@@ -322,7 +360,7 @@ SEXP vec_unique_names(SEXP x, bool quiet) {
       describe_repair(names, out);
     }
   } else {
-    out = PROTECT(as_unique_names(names, quiet));
+    out = PROTECT(vec_as_unique_names(names, quiet));
   }
 
   UNPROTECT(2);
@@ -459,7 +497,50 @@ SEXP set_rownames(SEXP x, SEXP names) {
 }
 
 
+enum name_repair_arg validate_name_repair(SEXP arg) {
+  if (!Rf_length(arg)) {
+    Rf_errorcall(R_NilValue, "`.name_repair` must be a string. See `?vctrs::vec_as_names`.");
+  }
+
+  arg = r_chr_get(arg, 0);
+
+  if (arg == strings_none) {
+    return name_repair_none;
+  }
+  if (arg == strings_minimal) {
+    return name_repair_minimal;
+  }
+  if (arg == strings_unique) {
+    return name_repair_unique;
+  }
+  if (arg == strings_universal) {
+    return name_repair_universal;
+  }
+  if (arg == strings_check_unique) {
+    return name_repair_check_unique;
+  }
+
+  Rf_errorcall(R_NilValue, "`.name_repair` can't be \"%s\". See `?vctrs::vec_as_names`.", CHAR(arg));
+}
+
+// [[ include("vctrs.h") ]]
+const char* name_repair_arg_as_c_string(enum name_repair_arg arg) {
+  switch (arg) {
+  case name_repair_none: return "none";
+  case name_repair_minimal: return "minimal";
+  case name_repair_unique: return "unique";
+  case name_repair_universal: return "universal";
+  case name_repair_check_unique: return "check_unique";
+  }
+}
+
+
 void vctrs_init_names(SEXP ns) {
   syms_set_rownames = Rf_install("set_rownames");
+  syms_as_universal_names = Rf_install("as_universal_names");
+  syms_validate_unique_names = Rf_install("validate_unique");
+
   fns_set_rownames = r_env_get(ns, syms_set_rownames);
+  fns_as_universal_names = r_env_get(ns, syms_as_universal_names);
+  fns_validate_unique_names = r_env_get(ns, syms_validate_unique_names);
 }