+++ NEWS.md
@@ -1,8 +1,11 @@
 # vctrs 0.1.0.9000
 
+* `vec_c()`, `vec_rbind()`, and `vec_cbind()` gain a `.name_repair`
+  argument (#227, #229).
+
 * `vec_dims()` has been replaced by `vec_dim_n()`. This name more clearly
   indicates the purpose of the function.
-  
+
 * New `stop_incompatible_size()` to signal a failure due to mismatched sizes.
 
 * `vec_empty()` has been renamed to `vec_is_empty()`.